hw14 fix

+++ src/samples/shadowmap/gui.cpp
@@ -14,7 +14,7 @@ void SimpleShadowmapRender::SetupGUIElements()
     ImGui::ColorEdit3("Meshes base color", m_uniforms.baseColor.M, ImGuiColorEditFlags_PickerHueWheel | ImGuiColorEditFlags_NoInputs);
     ImGui::SliderFloat3("Light source position", m_uniforms.lightPos.M, -10.f, 10.f);
     ImGui::SliderFloat3("Quad xz scale and y offset", pushConstQuad.scaleAndOffset.M, -10.f, 10.f);
-    ImGui::SliderFloat("Tessellation level", &pushConstQuad.tes_level, 1.f, 256.f);
+    ImGui::SliderInt("Tessellation level", &pushConstQuad.tes_level, 1, 256);
 
     if (ImGui::SliderFloat("Max height", &pushConstQuad.maxHeight, pushConstQuad.minHeight + 0.05f, 5.f))
       m_heightPass.recreate = true;
